requirements.txt
altair==4.2.0
Pillow==9.2.0
tensorflow==2.10.0
numpy==1.23.3
streamlit==1.13.0
opencv-python==4.6.0.66
plotly==5.10.0

